podofo: Add recipe for 1.1.2

+++ recipes/podofo/all/conanfile.py
@@ -26,6 +26,7 @@ class PodofoConan(ConanFile):
         "with_tiff": [True, False],
         "with_png": [True, False],
         "with_unistring": [True, False],
+        "with_fontconfig": [True, False],
     }
     default_options = {
         "shared": False,
@@ -35,6 +36,7 @@ class PodofoConan(ConanFile):
         "with_tiff": True,
         "with_png": True,
         "with_unistring": True,
+        "with_fontconfig": True,
     }
 
     def export_sources(self):
@@ -93,6 +95,9 @@ class PodofoConan(ConanFile):
         if self.options.with_unistring:
             self.requires("libunistring/0.9.10")
 
+        if self.options.with_fontconfig and Version(self.version) >= "1.1.2":
+            self.requires("fontconfig/2.15.0")
+
     def build_requirements(self):
         self.tool_requires("cmake/[>=3.16]")
 
@@ -103,20 +108,34 @@ class PodofoConan(ConanFile):
         get(self, **self.conan_data["sources"][self.version],
                   destination=self.source_folder, strip_root=True)
         apply_conandata_patches(self)
-        # Unvendor 3rd party libraries
-        libraries_to_unvendor = ["fmt", "date", "utf8cpp", "tclap"]
-        if Version(self.version) >= "1.0.3":
-            libraries_to_unvendor.append("utf8proc")
-            rm(self, "span.hpp", os.path.join(self.source_folder, "src", "podofo", "3rdparty"))
-        else:
-            rm(self, "span.hpp", os.path.join(self.source_folder, "src", "podofo", "auxiliary"))
-        for vendor_library in libraries_to_unvendor:
-            rmdir(self, os.path.join(self.source_folder, "3rdparty", vendor_library))
-        # Remove single header fast_float to use the conan package
-        rm(self, "fast_float.h", os.path.join(self.source_folder, "3rdparty"))
+
+        # Unvendor 3rd party libraries on versions where PoDoFo didn't support this directly
+        if Version(self.version) < "1.1.2":
+            libraries_to_unvendor = ["fmt", "date", "utf8cpp", "tclap"]
+            if Version(self.version) >= "1.0.3":
+                libraries_to_unvendor.append("utf8proc")
+                rm(self, "span.hpp", os.path.join(self.source_folder, "src", "podofo", "3rdparty"))
+            else:
+                rm(self, "span.hpp", os.path.join(self.source_folder, "src", "podofo", "auxiliary"))
+            for vendor_library in libraries_to_unvendor:
+                rmdir(self, os.path.join(self.source_folder, "3rdparty", vendor_library))
+            # Remove single header fast_float to use the conan package
+            rm(self, "fast_float.h", os.path.join(self.source_folder, "3rdparty"))
 
     def generate(self):
         tc = CMakeToolchain(self)
+        # Unvendor 3rd party libraries via native support by PoDoFo
+        if Version(self.version) >= "1.1.2":
+            tc.cache_variables["PODOFO_DEVENDOR_TCBSPAN"] = True
+            tc.cache_variables["PODOFO_DEVENDOR_DATE"] = True
+            tc.cache_variables["PODOFO_DEVENDOR_FASTFLOAT"] = True
+            tc.cache_variables["PODOFO_DEVENDOR_FMT"] = True
+            tc.cache_variables["PODOFO_DEVENDOR_UTF8CPP"] = True
+            tc.cache_variables["PODOFO_DEVENDOR_UTF8PROC"] = True
+
+            if not self.options.with_fontconfig:
+                tc.cache_variables["PODOFO_WITH_FONTMANAGER"] = False
+
         tc.cache_variables["PODOFO_BUILD_TEST"] = False
         tc.cache_variables["PODOFO_BUILD_EXAMPLES"] = False
         tc.cache_variables["PODOFO_BUILD_STATIC"] = not self.options.shared
